remove login function

+++ webapp/src/main/java/jerry/sideproject/web/webapp/controller/MultipleItemsController.java
@@ -1,9 +1,10 @@
 package jerry.sideproject.web.webapp.controller;
 
 import jerry.sideproject.web.webapp.controller.beans.ItemDto;
-import jerry.sideproject.web.webapp.controller.beans.ItemsCreationDto;
+import jerry.sideproject.web.webapp.controller.beans.ItemDtoList;
 import jerry.sideproject.web.webapp.services.interfaces.ItemService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -27,7 +28,7 @@ public class MultipleItemsController {
 
     @GetMapping(value = "/create")
     public String showCreateForm(Model model) {
-        ItemsCreationDto itemForm = new ItemsCreationDto();
+        ItemDtoList itemForm = new ItemDtoList();
         for (int i = 1; i <= 3; i++) {
             itemForm.addItem(new ItemDto());
         }
@@ -41,23 +42,25 @@ public class MultipleItemsController {
         itemService.findAll()
                 .iterator()
                 .forEachRemaining(itemDtos::add);
-        model.addAttribute("form", new ItemsCreationDto(itemDtos));
+        model.addAttribute("form", new ItemDtoList(itemDtos));
         return "editItemsForm";
     }
 
     @PostMapping(value = "/save")
-    public String saveItems(@ModelAttribute ItemsCreationDto form, Model model) {
-        itemService.saveAll(form.getItemDtos());
+    public String saveItems(@ModelAttribute ItemDtoList form, Model model) {
+        itemService.saveAll(form.getItems());
         model.addAttribute("items", itemService.findAll());
         return "redirect:/items/all";
     }
 
     @PostMapping(value = "/saveEdition")
-    public String saveEdition(@ModelAttribute ItemsCreationDto form, @RequestParam("deleteItems") Long[] deleteItems, Model model) {
-        itemService.saveAll(form.getItemDtos());
+    public String saveEdition(@ModelAttribute ItemDtoList form, @RequestParam("deleteItems") @Nullable Long[] deleteItems, Model model) {
+        itemService.saveAll(form.getItems());
         model.addAttribute("items", itemService.findAll());
-        for (Long id : deleteItems) {
-            itemService.remove(id);
+        if (deleteItems != null) {
+            for (Long id : deleteItems) {
+                itemService.remove(id);
+            }
         }
         return "redirect:/items/all";
     }
